Finally, KSCD uses newer WorkMan code. This means, that KSCD now needs libWorkMan (included in libwm/) to work. Because the code (both libworkman and kscd) is a little rough cut, I await lots of problems running kscd now. At least on platforms other than Linux and Solaris.
I hope to fix this ASAP, especially after removing the compile time
configuration.

Be prepared to see lots of output on stderr with this version of kscd.
This comes from the libWorkMan diagnostics and is run-time configurable
(but I didn't implement it in kscd so far).

Please forgive me my incomplete work in the makefiles, but I've got no
time left and had to commit.

svn path=/trunk/kdemultimedia/kscd/libwm/; revision=51207

+++ include/wm_cddb.h
@@ -31,7 +31,7 @@ extern int cur_cddb_protocol;
 extern char *cur_cddb_server, *cur_cddb_mail_adress,
 	*cur_cddb_path_to_cgi, *cur_cddb_proxy_server;
 
-unsigned int cddb_discid();
+unsigned long cddb_discid();
 void cddb_struct2cur();
 void cddb_cur2struct();
 void cddb_select();
